+++ surf-shop-core/surf-shop-core-common/src/main/kotlin/dev/slne/surf/shop/core/common/service/ShopService.kt
@@ -13,7 +13,7 @@ interface ShopService {
     suspend fun createShop(
         itemString: String,
         storedItemCount: Int,
-        pricePerItem: Int,
+        pricePerItem: Double,
         seller: UUID
     ): Shop
